% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{AncBetaOut}
\alias{AncBetaOut}
\title{Ancestry Generator}
\usage{
AncBetaOut(mean, var, n)
}
\arguments{
\item{mean}{mean of ancestry}

\item{n}{number of individuals}

\item{sd}{sd of ancestry}
}
\value{
A matrix of global ancestry percentages for n individuals
}
\description{
Helper function to generate 2-way admixture under beta distribution.
}
\examples{
AncBetaOut(0.75, 0.01, 100)
}
