% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_anc.R
\name{powphenanc}
\alias{powphenanc}
\title{Power Estimate for Trait Ancestry Associations}
\usage{
powphenanc(
  incidence.rates,
  N.adm = 3e+05,
  anc.matrix,
  n = 1000,
  case.size = 750,
  control.size = 750,
  h2 = 1,
  liability = NA,
  anc.names = NA
)
}
\arguments{
\item{incidence.rates}{A vector of incidence rates for the ancestral populations.}

\item{N.adm}{Size of simulated individuals from which cases and controls are drawn from. Default is 3e5.}

\item{anc.matrix}{A sample matrix of global ancestries, where rows represent samples and each column is one ancestry. See the format from the output of \code{AncBetaOut} function. If an ancestry matrix from empirical estimates is not available, \code{AncBetaOut} can be used to supply ancestry draws from simulations.}

\item{n}{Iteration number of phenotype-ancestry regression tests. Each iteration is an independent simulation of phenotypes and ancestries for case and controls. Default 1000.}

\item{case.size}{Number of cases. Default 750.}

\item{control.size}{Number of controls. Default 750.}

\item{h2}{Heritability. Default 1.}

\item{liability}{Liability of the binary trait, used for defining cases and controls. If left as NA (default), then a proxy phenotype based on ancestry and incidence rates is used to reflect the probability of becoming cases.}

\item{anc.names}{A vector of names for each ancestry, e.g. South Asian, West African, etc. It can be left unspecified (NA as default).}
}
\value{
A list of the following components:\tabular{ll}{
\code{ancestry} \tab Name of ancestries\cr
\tab \cr
\code{liability_model} \tab If a liability treshold model has been used\cr
\tab \cr
\code{powers} \tab Power per ancestry\cr
\tab \cr}
}
\description{
Estimates power for significant associations between a binary trait and global ancestries.
}
\examples{
# sample ancestries for a size of 100
aa_anc <- AncBetaOut(0.75, 0.01, 100)\cr
# simulate phenotype-ancestry associations
aa_ancp <- powphenanc(incidence.rates=c(0.01, 0.05), anc.matrix=aa_anc, h2=0.5, anc.names=c("eur","waf"))\cr
# check powers
anc_powers <- aa_ancp$powers\cr
}
\seealso{
\code{AncBetaOut}
}
