% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_geno.R
\name{SimGenoPower}
\alias{SimGenoPower}
\title{Power Inference for GWAS Discovery in Admixed Populations}
\usage{
SimGenoPower(
  n_snps = 1000,
  popsizes = c(1000, 1000, 1000),
  mean_ganc = 0.75,
  sd_ganc = 0.1,
  prop_causal = 0.1,
  hg2 = 0.5,
  fst = 0.2,
  fst_constant = F,
  freqdist = "unif",
  liability = NA,
  aaf = c(0.001, 0.999),
  envanc = 1,
  envvar = 0,
  alpha = 0.05
)
}
\arguments{
\item{n_snps}{Total number of SNPs (causal and non-causal). SNPs are assumed independent from each other. Default is 1000.}

\item{popsizes}{A vector of sample sizes for the admixed group, Pop 1 (ancestral origin from which the major component of ancestry comes), Pop 2 (ancestral origin from which the minor component of ancestry comes). Default is 1000 each.}

\item{mean_ganc}{Mean global ancestry from Pop 1 in the admixed group. Default is 0.75.}

\item{sd_ganc}{Standard deviation of the global ancestry distribution from Pop 1 in the admixed group. Default is 0.1.}

\item{prop_causal}{Proportion of causal variant among total SNPs. Default is 0.1.}

\item{hg2}{Heritability of the trait. Default is 0.5.}

\item{fst}{The baseline fixation index FST among causal variants between Pop 1 and Pop2. Default is 0.2.}

\item{fst_constant}{True or False: Is FST among causal variants a constant value equaling the baseline (True) or have an increment with the rarity of ancestral minor allele frequency (False)? Default is False.}

\item{freqdist}{Frequency distribution of ancestral variants (prior to the divergence of Pop 1 and 2). "unif" for a uniform distribution modeling (default), "inverse" for an inverse distribution modeling.}

\item{liability}{A vector of liability threshold for dichotomous traits in the admixed group, Pop 1, Pop2, respectively. If set as NA (default), the trait is assumed to be quantitative.}

\item{aaf}{Limits (min, max) on the frequency distribution of ancestral variants. Default is c(0.001, 0.999).}

\item{envanc}{Options to model environment by ancestry interactions: 0 means no interaction. 1 means interation is modeled as the sum of Gaussian sampling. 2 means the interaction is modeled as linear dependence on the ancestry proportion. Default is 1.}

\item{envvar}{percentage of the environment by ancestry interaction over the total phenotypic variance. Only effective when \code{envanc} is 2.}

\item{alpha}{Type I error rate. Default is 0.05.}
}
\value{
A list of powers and the re-estimated heritability from simulation.\tabular{ll}{
\code{admixed_power} \tab Power for discovery in the admixed group\cr
\tab \cr
\code{pop1_power} \tab Power for discovery in Pop 1\cr
\tab \cr
\code{pop2_power} \tab Power for discovery in Pop 2\cr
\tab \cr
\code{h2} \tab Estimated heritability from simulated phenotypes and genotypes \cr
\tab \cr}
}
\description{
Infers statistical power for discovery of variants in a 2-way admixed group, and the corresponding populations of ancestry origin.
}
